jwt security added
