+++ srv/user-service.cds
@@ -1,7 +1,7 @@
 using { cap_tutorial as my } from '../db/schema';
 namespace cap_tutorial; 
 
-service UsersService  { 
+service UsersService { 
   entity Users as projection on my.Users;
 
 }
